requirements.txt
streamlit==1.42.2
openai-whisper==20231117
torch>=2.1,<3.0
numpy>=1.26,<3.0
pandas>=2.1,<3.0
soundfile>=0.12.1
plotly>=5.18
python-docx>=1.1.0
openpyxl>=3.1.2
imageio-ffmpeg>=0.6.0
